// src/app/models/student.ts
export class Student {
    MaHS: number;
    TenHS: string;
    NgaySinh: string;
    GioiTinh: string;
    TiengAnh: number;
    CongNghe: number;
    GiaoDuc: number;
    HoaHoc: number;
    LichSu: number;
    SinhHoc: number;
    Toan: number;
    TinHoc: number;
    Van:number;
    DiemTB: number;
}